+++ src/reports/smtp_tls.rs
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-use actix_web::{web::{Data, Json}, HttpResponse, Responder};
+use actix_web::{http::header, web::{Data, Json}, HttpRequest, HttpResponse, Responder};
 use log::error;
 use serde::{Deserialize, Serialize};
 
@@ -90,8 +90,12 @@ pub struct SMTPTLSReport {
     policies: Vec<PoliciesItem>
 }
 
-pub async fn report_smtp_tls(state: Data<WebState>, report: Json<SMTPTLSReport>) -> impl Responder {
-    let res = handle_report(&ReportType::SMTPTLSRPT(&report), &state.filter).await;
+pub async fn report_smtp_tls(state: Data<WebState>, req: HttpRequest, report: Json<SMTPTLSReport>) -> impl Responder {
+    let res = handle_report(
+        &ReportType::SMTPTLSRPT(&report), 
+        req.headers().get(header::USER_AGENT).map(|h| h.to_str().unwrap()),
+        &state.filter
+    ).await;
     match res {
         Ok(_) => HttpResponse::Ok(),
         Err(err) => {