+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/nerou42/network-journal/compare/v0.3.0...HEAD)
 
+### Added
+
+- Support for the following types of reports:
+    - DMARC
+
 
 ## [0.3.0](https://github.com/nerou42/network-journal/compare/v0.2.0...v0.3.0)