+++ CHANGELOG.md
@@ -5,12 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/nerou42/network-journal/compare/v0.3.0...HEAD)
+## [Unreleased](https://github.com/nerou42/network-journal/compare/v0.4.0...HEAD)
+
+
+## [0.4.0](https://github.com/nerou42/network-journal/compare/v0.3.0...v0.4.0)
 
 ### Added
 
 - Support for the following types of reports:
     - DMARC
+
+### Changed
+
+- Improve config file change handling by implementing default
 
 
 ## [0.3.0](https://github.com/nerou42/network-journal/compare/v0.2.0...v0.3.0)